+++ frontend/src/app/policy-engine/policy-configuration/blocks/calculate/custom-logic-config/custom-logic-config.component.ts
@@ -3,7 +3,7 @@ import { BlockNode } from '../../../../helpers/tree-data-source/tree-data-source
 import { MatDialog } from '@angular/material/dialog';
 import { CodeEditorDialogComponent } from '../../../../helpers/code-editor-dialog/code-editor-dialog.component';
 import { Schema, Token, SchemaField } from '@guardian/interfaces';
-import { PolicyBlockModel, PolicyModel } from 'src/app/policy-engine/policy-model';
+import { PolicyBlockModel, PolicyModel } from 'src/app/policy-engine/structures/policy-model';
 
 @Component({
   selector: 'app-custom-logic-config',
