+++ frontend/src/app/policy-engine/policy-configuration/blocks/documents/external-data-config/external-data-config.component.ts
@@ -1,6 +1,6 @@
 import { Component, EventEmitter, Input, OnInit, Output, SimpleChanges } from '@angular/core';
 import { Schema, Token } from '@guardian/interfaces';
-import { PolicyBlockModel, PolicyModel } from 'src/app/policy-engine/policy-model';
+import { PolicyBlockModel, PolicyModel } from 'src/app/policy-engine/structures/policy-model';
 import { BlockNode } from '../../../../helpers/tree-data-source/tree-data-source';
 
 /**
