+++ frontend/src/app/policy-engine/policy-configuration/blocks/documents/revoke-config/revoke-config.component.ts
@@ -1,6 +1,6 @@
 import { Component, EventEmitter, Input, OnInit, Output, SimpleChanges } from '@angular/core';
 import { Schema, Token } from '@guardian/interfaces';
-import { PolicyBlockModel } from 'src/app/policy-engine/policy-model';
+import { PolicyBlockModel } from 'src/app/policy-engine/structures/policy-model';
 import { BlockNode } from '../../../../helpers/tree-data-source/tree-data-source';
 
 /**
