+++ frontend/src/app/policy-engine/policy-configuration/blocks/tokens/token-action-config/token-action-config.component.ts
@@ -1,6 +1,6 @@
 import { Component, EventEmitter, Input, OnInit, Output, SimpleChanges } from '@angular/core';
 import { Schema, Token } from '@guardian/interfaces';
-import { PolicyBlockModel, PolicyModel } from 'src/app/policy-engine/policy-model';
+import { PolicyBlockModel, PolicyModel } from 'src/app/policy-engine/structures/policy-model';
 /**
  * Settings for block of 'mintDocument' and 'wipeDocument' types.
  */
