+++ frontend/src/app/policy-engine/policy-configuration/policy-configuration/policy-configuration.component.ts
@@ -7,15 +7,16 @@ import { forkJoin, Observable } from 'rxjs';
 import { ActivatedRoute, Router } from '@angular/router';
 import { MatDialog } from '@angular/material/dialog';
 import { TokenService } from 'src/app/services/token.service';
-import { BlockGroup, RegisteredBlocks } from '../../registered-blocks';
+import { RegisteredBlocks } from '../../registered-blocks';
+import { BlockGroup } from "../../structures/types/block-group.type";
 import { PolicyAction, SavePolicyDialog } from '../../helpers/save-policy-dialog/save-policy-dialog.component';
 import { SetVersionDialog } from 'src/app/schema-engine/set-version-dialog/set-version-dialog.component';
 import * as yaml from 'js-yaml';
 import { Clipboard } from '@angular/cdk/clipboard';
 import { ConfirmationDialogComponent } from 'src/app/components/confirmation-dialog/confirmation-dialog.component';
 import { EventsOverview } from '../../helpers/events-overview/events-overview';
-import { PolicyBlockModel, PolicyModel } from '../../policy-model';
-import { PolicyStorage } from '../../policy-storage';
+import { PolicyBlockModel, PolicyModel } from '../../structures/policy-model';
+import { PolicyStorage } from '../../structures/policy-storage';
 import { TreeFlatOverview } from '../../helpers/tree-flat-overview/tree-flat-overview';
 import { SaveBeforeDialogComponent } from '../../helpers/save-before-dialog/save-before-dialog.component';
 import { TasksService } from 'src/app/services/tasks.service';
