+++ frontend/src/app/policy-engine/structures/interfaces/children-display-settings.interface.ts
@@ -1,0 +1,9 @@
+import { BlockGroup } from "../types/block-group.type";
+import { BlockHeaders } from "../types/block-headers.type";
+import { BlockType } from "../types/block-type.type";
+
+export interface ChildrenDisplaySettings {
+    type: BlockType;
+    group?: BlockGroup;
+    header?: BlockHeaders;
+}
