+++ frontend/src/app/policy-engine/structures/types/block-group.type.ts
@@ -1,0 +1,9 @@
+
+export enum BlockGroup {
+    Main = 'Main',
+    Documents = 'Documents',
+    Tokens = 'Tokens',
+    Calculate = 'Calculate',
+    Report = 'Report',
+    UnGrouped = 'UnGrouped'
+}
