+++ frontend/src/app/policy-engine/structures/types/block-headers.type.ts
@@ -1,0 +1,6 @@
+
+export enum BlockHeaders {
+    UIComponents = 'UI Components',
+    ServerBlocks = 'Server Blocks',
+    Addons = 'Addons'
+}
