+++ frontend/src/app/policy-engine/structures/types/block-type.type.ts
@@ -1,0 +1,34 @@
+
+export enum BlockType {
+    Container = 'interfaceContainerBlock',
+    DocumentsViewer = 'interfaceDocumentsSourceBlock',
+    Information = 'informationBlock',
+    PolicyRoles = 'policyRolesBlock',
+    Request = 'requestVcDocumentBlock',
+    SendToGuardian = 'sendToGuardianBlock',
+    Action = 'interfaceActionBlock',
+    Step = 'interfaceStepBlock',
+    Mint = 'mintDocumentBlock',
+    ExternalData = 'externalDataBlock',
+    AggregateDocument = 'aggregateDocumentBlock',
+    Wipe = 'retirementDocumentBlock',
+    FiltersAddon = 'filtersAddon',
+    DocumentsSourceAddon = 'documentsSourceAddon',
+    Calculate = 'calculateContainerBlock',
+    CalculateMathAddon = 'calculateMathAddon',
+    Report = 'reportBlock',
+    ReportItem = 'reportItemBlock',
+    ReassigningBlock = 'reassigningBlock',
+    PaginationAddon = 'paginationAddon',
+    TimerBlock = 'timerBlock',
+    CustomLogicBlock = 'customLogicBlock',
+    Switch = 'switchBlock',
+    RevokeBlock = 'revokeBlock',
+    SetRelationshipsBlock = 'setRelationshipsBlock',
+    ButtonBlock = 'buttonBlock',
+    TokenActionBlock = 'tokenActionBlock',
+    DocumentValidatorBlock = 'documentValidatorBlock',
+    TokenConfirmationBlock = 'tokenConfirmationBlock',
+    GroupManagerBlock = 'groupManagerBlock',
+    MultiSigBlock = 'multiSigBlock'
+}
