+++ frontend/src/app/policy-engine/structures/types/children-type.type.ts
@@ -1,0 +1,6 @@
+
+export enum ChildrenType {
+    None = 'None',
+    Special = 'Special',
+    Any = 'Any'
+}
