+++ frontend/src/app/policy-engine/structures/types/control-type.type.ts
@@ -1,0 +1,7 @@
+
+export enum ControlType {
+    UI = 'UI',
+    Special = 'Special',
+    Server = 'Server',
+    None = 'None'
+}
