+++ guardian-service/src/policy-engine/blocks/index.ts
@@ -30,3 +30,4 @@ export { TokenActionBlock } from './token-action-block';
 export { DocumentValidatorBlock } from './document-validator-block';
 export { TokenConfirmationBlock } from './token-confirmation-block';
 export { GroupManagerBlock } from './group-manager';
+export { MultiSigBlock } from './multi-sig-block';
