+++ guardian-service/src/policy-engine/blocks/multi-sig-block.ts
@@ -1,0 +1,60 @@
+import { ActionCallback, BasicBlock } from '@policy-engine/helpers/decorators';
+import { PolicyValidationResultsContainer } from '@policy-engine/policy-validation-results-container';
+import { PolicyComponentsUtils } from '@policy-engine/policy-components-utils';
+import { PolicyUtils } from '@policy-engine/helpers/utils';
+import { IPolicyEvent, PolicyInputEventType, PolicyOutputEventType } from '@policy-engine/interfaces';
+import { ChildrenType, ControlType, PropertyType } from '@policy-engine/interfaces/block-about';
+import { IPolicyDocument, IPolicyEventState } from '@policy-engine/policy-engine.interface';
+
+/**
+ * Switch block
+ */
+@BasicBlock({
+    blockType: 'multiSigBlock',
+    commonBlock: true,
+    about: {
+        label: 'Switch',
+        title: `Add 'Switch' Block`,
+        post: false,
+        get: false,
+        children: ChildrenType.None,
+        control: ControlType.Server,
+        input: [
+            PolicyInputEventType.RunEvent
+        ],
+        output: [
+            PolicyOutputEventType.RefreshEvent
+        ],
+        defaultEvent: false,
+        properties: [{
+            name: 'threshold',
+            label: 'Threshold (%)',
+            type: PropertyType.Input
+        }]
+    }
+})
+export class MultiSigBlock {
+    /**
+     * Run block action
+     * @event PolicyEventType.Run
+     * @param {IPolicyEvent} event
+     */
+    @ActionCallback({
+        output: [PolicyOutputEventType.RunEvent, PolicyOutputEventType.RefreshEvent]
+    })
+    async runAction(event: IPolicyEvent<IPolicyEventState>) {
+        const ref = PolicyComponentsUtils.GetBlockRef(this);
+    }
+
+    /**
+     * Validate block options
+     * @param resultsContainer
+     */
+    public async validate(resultsContainer: PolicyValidationResultsContainer): Promise<void> {
+        const ref = PolicyComponentsUtils.GetBlockRef(this);
+        try {
+        } catch (error) {
+            resultsContainer.addBlockError(ref.uuid, `Unhandled exception ${PolicyUtils.getErrorMessage(error)}`);
+        }
+    }
+}
