+++ guardian-service/src/policy-engine/interfaces/block-about.ts
@@ -19,6 +19,81 @@ export enum ControlType {
     None = 'None',
 }
 
+export enum PropertyType {
+    Input = 'Input',
+    Checkbox = 'Checkbox',
+    Select = 'Select',
+    MultipleSelect = 'MultipleSelect',
+    Group = 'Group',
+    Array = 'Array'
+}
+
+/**
+ * Block Properties
+ */
+export interface BlockProperties {
+    name: string;
+    label: string;
+    type: PropertyType;
+}
+
+/**
+ * Input Properties
+ */
+export interface InputProperties extends BlockProperties {
+    type: PropertyType.Input;
+}
+
+/**
+ * Checkbox Properties
+ */
+export interface CheckboxProperties extends BlockProperties {
+    type: PropertyType.Checkbox;
+
+}
+
+/**
+ * Select Properties
+ */
+export interface SelectProperties extends BlockProperties {
+    type: PropertyType.Select;
+
+}
+
+/**
+ * MultipleSelect Properties
+ */
+export interface MultipleSelectProperties extends BlockProperties {
+    type: PropertyType.MultipleSelect;
+
+}
+
+/**
+ * Group Properties
+ */
+export interface GroupProperties extends BlockProperties {
+    type: PropertyType.Group;
+    properties?: AnyBlockProperties[];
+}
+
+/**
+ * Array Properties
+ */
+export interface ArrayProperties extends BlockProperties {
+    type: PropertyType.Array;
+}
+
+/**
+ * Any Block Properties
+ */
+export type AnyBlockProperties =
+    InputProperties |
+    CheckboxProperties |
+    SelectProperties |
+    MultipleSelectProperties |
+    GroupProperties |
+    ArrayProperties;
+
 /**
  * Block about
  */
@@ -59,4 +134,8 @@ export interface BlockAbout {
      * Default event
      */
     defaultEvent: boolean;
+    /**
+     * Default properties
+     */
+    properties?: AnyBlockProperties[];
 }
